Reduce number of types files
