--- src/envalidWithDotenv.d.ts
@@ -1,1 +1,0 @@
-export * from './envalid'
