+++ src/envalidWithoutDotenv.d.ts
@@ -1,0 +1,3 @@
+// This file is used for browsers and react-native.
+// The dotEnvPath option will be ignored.
+export * from './envalid'
