--- src/indexNoDotenv.d.ts
@@ -1,2 +1,0 @@
-// NOTE: dotEnvPath option will be ignored when using this file.
-export * from './envalid'
