--- src/indexNoDotenv.js
@@ -1,4 +1,0 @@
-// NOTE: dotEnvPath option will be ignored when using this file.
-// This file is used to support non-Node environments, such as
-// browsers and React Native.
-module.exports = require('./envalid')
